% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dec.sim.R
\name{plot.dec.sim}
\alias{plot.dec.sim}
\title{plot simulation results from a dec.sim object}
\usage{
\method{plot}{dec.sim}(x, pt, s = 1, type = c("all", "s", "prob", "np",
  "dlt"), show = TRUE, col = "cornflowerblue", text.col = "darkblue",
  cex = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"dec.sim"} or \code{"sl.sim"}, a result of a call to \code{dec.sim} or \code{sl.sim}.}

\item{pt}{a vector with target toxicity for each scenario.}

\item{s}{scenario to be plotted. Defaults to 1.}

\item{type}{plot type. See descriptions above.}

\item{show}{a logical value indicating if values are shown on plot.}

\item{col}{graphical parameter \code{col}; see details \code{\link{par}}.}

\item{text.col}{plotting color of text shown.}

\item{cex}{graphical parameter \code{col}; see details \code{\link{par}}.}

\item{...}{arguments to be passed to \code{\link{plot}} methods.}
}
\description{
Three plots are currently available: a plot of true toxicity at each dose level (\code{type = "s"}); a bar plot of the probability of selecting as the MTD for each dose level (\code{type = "prob"}); a bar plot of the average number of patients treated at each dose level (\code{type = "np"}); a bar plot of the average number of patients experienced DLT at each dose level (\code{type = "dlt"}) and \code{type = "all"} generates all above plots.
}
