% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dec.table.R
\name{plot.dec.table}
\alias{plot.dec.table}
\title{plot decision table from a "dec.table" object.}
\usage{
\method{plot}{dec.table}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"dec.table"}, a result of a call to \code{dec.table}.}

\item{...}{Not used argument.}
}
\description{
\code{plot} method for class "\code{dec.table}"
}
\details{
\code{plot.dec.table} prints the decision boundarys.
}
