% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec.table.R
\name{print.dec.table}
\alias{print.dec.table}
\title{print decision table from a "dec.table" object.}
\usage{
\method{print}{dec.table}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"dec.table"}, a result of a call to \code{dec.table}.}

\item{...}{Not used argument.}
}
\description{
\code{print} method for class "\code{dec.table}"
}
\details{
\code{print.dec.table} prints the decision table with legend keys.
}
