% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.opt.design}
\alias{print.opt.design}
\title{print Zhong's design from a "opt.design" object.}
\usage{
\method{print}{opt.design}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"opt.design"}, a result of a call to \code{opt.design}.}

\item{...}{not used argument.}
}
\description{
\code{print} method for class "\code{opt.design}"
}
