% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dec.sim.R
\name{summary.dec.sim}
\alias{summary.dec.sim}
\title{Summarizing simulation results from a dec.sim object}
\usage{
\method{summary}{dec.sim}(object, pt, ...)
}
\arguments{
\item{object}{an object of class \code{"dec.sim"}, a result of a call to \code{dec.sim} or \code{sl.sim}.}

\item{pt}{target toxicity for each scenario.}

\item{...}{Not used argument.}
}
\description{
\code{summary} method for class \code{"dec.sim"}.
}
