% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.design.R
\name{opt.design}
\alias{opt.design}
\title{Zhong's 2-/3- stage Phase II design}
\usage{
opt.design(alpha1, alpha2, beta, pc, pt, stage = 2, frac_n1 = c(0.2, 0.3),
  frac_n2 = c(0.2, 0.4), sf = "Pocock", show = FALSE, ...)
}
\arguments{
\item{alpha1}{left-side overall type I error.}

\item{alpha2}{right-side overall type I error.}

\item{beta}{type II error}

\item{pc}{a numeric vector of response rate. Should be a vector with 1 or 2.}

\item{pt}{alternative hypothesis}

\item{stage}{2 or 3. default to 2 (2-stage design).}

\item{frac_n1}{proportion of n1. Used for 3-stage design. default to \code{c(0.2, 0.3)}, i.e. the range of \code{n1} is \code{0.2*n} to \code{0.3*n}}

\item{frac_n2}{proportion of n2. Used for 3-stage design. default to \code{c(0.2, 0.4)}.}

\item{sf}{the alpha-spending function to be used. \code{sf="OF"} or "\code{sf="Pocock"} uses spending function in R package \code{\link{gsDesign}}; or a user supplied spending function.}

\item{show}{logical. If \code{TRUE}, current sample size is shown as total sample size increase.}

\item{...}{not used argument.}
}
\value{
An object of class "opt.design" is a list containing:
 \item{bdry}{rejection regions}
 \item{error}{true type 1/2 errors}
 \item{n}{sample size at each stage} 
 \item{alpha1}{input; left-side type 1 error.}
 \item{alpha2}{input; right-side type 1 error.}
 \item{pc}{input; a vector of response rate.}
 \item{pt}{input; a vector of alternative response rate.}
 \item{sf}{input; the alpha-spending function used.}
 \item{stage}{input; two or three stage}
}
\description{
calculate optimal 2-/3-stage design given by Bob Zhong
}
\examples{
\dontrun{
 alpha1 <- 0.15
 alpha2 <- 0.10
 beta <- 0.15
 pc <- 0.25
 pt <- pc + 0.20
 opt.design(alpha1, alpha2, beta, pc, pt, stage=2, show=1)
 opt.design(alpha1, alpha2, beta, pc, pt, stage=3, show=1)
}
}
\references{
Zhong. (2012) Single-arm Phase IIA clinical trials with go/no-go decisions. \emph{Contemporary Clinical Trials}, \bold{33}, 1272--1279.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>
}
