% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dec.sim.R
\name{plot.dec.sim}
\alias{plot.dec.sim}
\title{plot simulation results from a dec.sim object}
\usage{
\method{plot}{dec.sim}(x, pt, s = 1, type = c("s", "prob", "np"),
  show = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{"dec.sim"}, a result of a call to \code{dec.sim} or \code{sl.sim}.}

\item{pt}{target toxicity for each scenario.}

\item{s}{scenario to be plotted. Default is 1.}

\item{type}{plot type. See descriptions above.}

\item{show}{a logical value indicating if values are shown on plot.}

\item{...}{Not used argument.}
}
\description{
Three plots are currently available: a plot of true toxicity at each dose level (\code{type="s"}); a bar plot of the probability of selecting as the MTD for each dose level (\code{type="prob"}) and a bar plot of the average number of patients treated at each dose level (\code{type="np"}).
}
