% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.opt.design}
\alias{print.opt.design}
\title{print Zhong's design from a "opt.design" object.}
\usage{
\method{print}{opt.design}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"opt.design"}, a result of a call to \code{opt.design}.}

\item{...}{not used argument.}
}
\description{
\code{print} method for class "\code{opt.design}"
}
\examples{
alpha1 <- 0.20
alpha2 <- 0.1
beta <- 0.20
pc <- 0.5
pt <- pc + 0.2
out <- opt.design(alpha1, alpha2, beta, pc, pt, stage = 2, sf.param = 1)
print(out)
}
